LPS-113924 moved getCKEditorReadOnlyConfig to utils.es.js
